Fix memory leak caused by non-unregistered broadcast receiver
