+++ android/app/build.gradle
@@ -63,6 +63,8 @@ dependencies {
     implementation "com.google.guava:guava:31.0.1-android"
     implementation 'com.google.code.gson:gson:2.8.6'
 
+    debugImplementation 'com.squareup.leakcanary:leakcanary-android:2.10'
+
     implementation "com.google.dagger:hilt-android:2.38.1"
     kapt "com.google.dagger:hilt-compiler:2.38.1"
