+++ android/app/src/main/java/app/candash/cluster/FullscreenActivity.kt
@@ -31,8 +31,8 @@ class FullscreenActivity : AppCompatActivity() {
     private var runnable: Runnable? = null
     private var delay = 1000
 
-
     private lateinit var viewModel: DashViewModel
+    private lateinit var hotSpotReceiver: HotSpotReceiver
 
     @RequiresApi(Build.VERSION_CODES.N)
     override fun onConfigurationChanged(newConfig: Configuration) {
@@ -59,22 +59,6 @@ class FullscreenActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         val prefs = getSharedPreferences("dash", Context.MODE_PRIVATE)
         super.onCreate(savedInstanceState)
-        val hotSpotReceiver = object : BroadcastReceiver() {
-            override fun onReceive(contxt: Context, intent: Intent) {
-                val action = intent.action
-                if ("android.net.wifi.WIFI_AP_STATE_CHANGED" == action) {
-                    val state = intent.getIntExtra(WifiManager.EXTRA_WIFI_STATE, 0)
-                    val prevState = intent.getIntExtra(WifiManager.EXTRA_PREVIOUS_WIFI_STATE, 0)
-                    if ((WifiManager.WIFI_STATE_ENABLED == state % 10) && (WifiManager.WIFI_STATE_ENABLED != prevState % 10)) {
-                        viewModel.restart()
-                    }
-                }
-            }
-        }
-        this.registerReceiver(
-            hotSpotReceiver,
-            IntentFilter("android.net.wifi.WIFI_AP_STATE_CHANGED")
-        )
         // check every second if battery is connected
         val context = applicationContext
 
@@ -142,6 +126,11 @@ class FullscreenActivity : AppCompatActivity() {
                 else -> throw IllegalStateException("Attempting to switch to unknown fragment: $it")
             }
         }
+        hotSpotReceiver = HotSpotReceiver(viewModel)
+        this.registerReceiver(
+            hotSpotReceiver,
+            IntentFilter("android.net.wifi.WIFI_AP_STATE_CHANGED")
+        )
     }
 
     private fun switchToFragment(fragment: Fragment) {
@@ -169,6 +158,7 @@ class FullscreenActivity : AppCompatActivity() {
 
     override fun onDestroy() {
         handler.removeCallbacks(runnable!!)
+        this.unregisterReceiver(hotSpotReceiver)
         super.onDestroy()
     }
 }
